Refactor to prepare RSA adding
