+++ example/App.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import Crypto from '@trackforce/react-native-crypto';
+import Aes from '@trackforce/react-native-crypto';
 declare var console: any;
 
 export default class App extends Component {
@@ -9,17 +9,17 @@ export default class App extends Component {
     try {
       const iterations = 4096;
       const keyInBytes = 32;
-      const key = await Crypto.pbkdf2('a0', 'a1b4efst', iterations, keyInBytes, 'SHA1');
+      const key = await Aes.pbkdf2('a0', 'a1b4efst', iterations, keyInBytes, 'SHA1');
       console.log(`pbkdf2 key: ${key}`);
 
       const iv = null; // or base 64 encoded 16 bytes random string
-      const ciphertext = await Crypto.aesEncrypt('data to encrypt', key, iv);
+      const ciphertext = await Aes.aesEncrypt('data to encrypt', key, iv);
       console.log(`aesEncrypt: ${ciphertext}`);
 
-      const decryptedText = await Crypto.aesDecrypt(ciphertext, key, iv);
+      const decryptedText = await Aes.aesDecrypt(ciphertext, key, iv);
       console.log(`aesDecrypt: ${decryptedText}`);
 
-      const hash = await Crypto.hmac256(ciphertext, key);
+      const hash = await Aes.hmac256(ciphertext, key);
       console.log(`hmac256: ${hash}`);
     } catch (err) {
       console.log('err', err);
