+++ example/android/app/src/main/java/com/example/MainApplication.java
@@ -3,7 +3,7 @@ package com.example;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.trackforce.aes.RCTCryptoPackage;
+import com.trackforce.crypto.RCTAesPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -25,7 +25,7 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-          new RCTCryptoPackage()
+          new RCTAesPackage()
       );
     }
