+++ example/package.json
@@ -9,11 +9,11 @@
     "test": "jest"
   },
   "dependencies": {
+    "@trackforce/react-native-crypto": "file:..",
     "react": "16.4.1",
     "react-native": "0.55.4",
-    "tslib": "^1.9.3",
-    "@trackforce/react-native-crypto": "file:../"
-    },
+    "tslib": "^1.9.3"
+  },
   "devDependencies": {
     "babel-preset-react-native": "^4.0.0",
     "react-native-typescript-transformer": "^1.2.10",
