+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 import { NativeModules } from 'react-native';
-export default NativeModules.Crypto;
+export default NativeModules.Aes;
