+++ ios/RCTCrypto/RCTAes.h
@@ -1,0 +1,5 @@
+#import <React/RCTBridgeModule.h>
+
+@interface RCTAes : NSObject <RCTBridgeModule>
+
+@end
