+++ ios/RCTCrypto/RCTAes.m
@@ -1,7 +1,7 @@
-#import "RCTCrypto.h"
-#import "AesCrypt.h"
+#import "RCTAes.h"
+#import "c.h"
 
-@implementation RCTCrypto
+@implementation RCTAes
 
 RCT_EXPORT_MODULE()
 
@@ -9,7 +9,7 @@ RCT_EXPORT_METHOD(aesEncrypt:(NSString *)data key:(NSString *)key iv:(NSString *
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *base64 = [AesCrypt aesEncrypt:data key:key iv:iv];
+    NSString *base64 = [Aes aesEncrypt:data key:key iv:iv];
     if (base64 == nil) {
         reject(@"encrypt_fail", @"Encrypt error", error);
     } else {
@@ -21,7 +21,7 @@ RCT_EXPORT_METHOD(aesDecrypt:(NSString *)base64 key:(NSString *)key iv:(NSString
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt aesDecrypt:base64 key:key iv:iv];
+    NSString *data = [Aes aesDecrypt:base64 key:key iv:iv];
     if (data == nil) {
         reject(@"decrypt_fail", @"Decrypt failed", error);
     } else {
@@ -33,7 +33,7 @@ RCT_EXPORT_METHOD(pbkdf2:(NSString *)password salt:(NSString *)salt iterations:(
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt pbkdf2:password salt:salt iterations:iterations keyLen:keyLen hash:hash];
+    NSString *data = [Aes pbkdf2:password salt:salt iterations:iterations keyLen:keyLen hash:hash];
     if (data == nil) {
         reject(@"keygen_fail", @"Key generation failed", error);
     } else {
@@ -45,7 +45,7 @@ RCT_EXPORT_METHOD(hmac256:(NSString *)base64 key:(NSString *)key
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt hmac256:base64 key:key];
+    NSString *data = [Aes hmac256:base64 key:key];
     if (data == nil) {
         reject(@"hmac_fail", @"HMAC error", error);
     } else {
@@ -57,7 +57,7 @@ RCT_EXPORT_METHOD(sha1:(NSString *)text
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt sha1:text];
+    NSString *data = [Aes sha1:text];
     if (data == nil) {
         reject(@"sha1_fail", @"Hash error", error);
     } else {
@@ -69,7 +69,7 @@ RCT_EXPORT_METHOD(sha256:(NSString *)text
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt sha256:text];
+    NSString *data = [Aes sha256:text];
     if (data == nil) {
         reject(@"sha256_fail", @"Hash error", error);
     } else {
@@ -81,7 +81,7 @@ RCT_EXPORT_METHOD(sha512:(NSString *)text
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     NSError *error = nil;
-    NSString *data = [AesCrypt sha512:text];
+    NSString *data = [Aes sha512:text];
     if (data == nil) {
         reject(@"sha512_fail", @"Hash error", error);
     } else {
