--- ios/RCTCrypto/RCTCrypto.h
@@ -1,5 +1,0 @@
-#import <React/RCTBridgeModule.h>
-
-@interface RCTCrypto : NSObject <RCTBridgeModule>
-
-@end
